Introduce package structure
This moves most of the functionality out of the zeek-client script and into
Python module called zeekclient, with submodules for the various
components. This changes nothing about the client's functionality, other than
minor renaming for a better fit in the (sub)module structure and avoidance of
cyclic references.

Also adds a .pylintrc.